third_party/blink/web_tests/external/wpt/html/semantics/popovers/popover-focus-2-expected.txt

This is a testharness.js-based test.
[FAIL] Popover focus navigation
  assert_equals: Focus should not move back to invoker as it is non-focusable expected Element node <button id="button1" tabindex="0">Button1</button> but got Element node <button popovertarget="popover0" id="invoker0" tabindex="...
Harness: the test ran to completion.
